related to type cast issue

+++ src/serugees-api/Models/ILoanRepository.cs
@@ -6,8 +6,8 @@ namespace serugees_api.Models
     {
         void Add(Loan item);
         IEnumerable<Loan> GetAll();
-        Loan Find(long key);
-        void Remove(long key);
+        Loan Find(int key);
+        void Remove(int key);
         void Update(Loan item);
     }
 }
